Create QuizGame.go
Basic quiz game intended for simple math questions. Input is a CSV file with questions and answers formatted as such:

2+2,4
3*5,15

default file is quiz_key.csv in the program folder, default time is 30 seconds to complete. Use -h for configuration options

This is my first program while teaching myself go. Some things are the way they are simply because I was exploring (Yes, I know, I opened the file twice and didn't need to. I was playing with os and io/ioutil packages)